LayoutTests/imported/w3c/web-platform-tests/svg/types/elements/SVGGeometryElement-rect-expected.txt


FAIL getTotalLength and getPointAtLength do not take pathLength into account assert_equals: expected 600 but got 0
FAIL getPointAtLength() returns instance of DOMPoint assert_true: expected true got false
